Text.Record.PrettyPrint

Documentation

class FQuery a

Accumulate Record accessors as [String] (in order of definition).

class VQuery a

Accumulate Record values as [String] (in order of definition).

class (Typeable a, FQuery a, Show a) => VFmt a

Reduce Record to Box given a Formatter

  Example:
     data List a = List { label :: String, val :: a, tail :: List a }
         | Nil deriving (Generic, Show)
     instance FQuery (List a)
     instance (Show a, Typeable a) => VFmt (List a)
     test = List "head" 3 $ List "mid" 4 $ List "tail" 5 $ Nil

     ghci> printBox $ format test dfltFmt
     label: "head"
     val: 3
     tail: label: "mid"
           val: 4
           tail: label: "tail"
                 val: 5
                 tail: Nil

format :: VFmt a => a -> Formatter a -> Box

fields :: FQuery a => a -> [String]

values :: VQuery a => a -> [String]

dfltFmt :: Formatter a

simpleFmt :: (VQuery a, FQuery a) => (String -> String -> b) -> a -> [b]

Given String representation of accessor and value return b

data Formatter a

A Formatter is responsible for converting a Record into a formatted Box.

   Example:
       dfltFmt :: Formatter a
       dfltFmt = Formatter
           { arg = text
           , label = \a b -> text (a <> ":") + b
           , finally = vcat left }

Constructors

Formatter

Fields

arg :: String -> Box

Converts Record fields to Boxes

label :: String -> Box -> Box

Merge accessor with field

finally :: [Box] -> Box

Finally reduce list of boxes